Fixed and refactored code
